pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "find_quantity"
description = "find_quantity help solve calcuation of quantity."
version = "1.0.8"
authors = [ {name = "Salah Eddine Saoud", email = "saoud92@gmail.com"} ]
license = {file = "LICENSE"}
dependencies = [
    'pyyaml',
]

[project.scripts]
find_quantity = "find_quantity.app:main"